requirements.txt
BASC-py4chan==0.6.5
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.1
future==0.18.2
geojson==2.5.0
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.8
markovify==0.8.0
mysqlclient==1.4.6
pkg-resources==0.0.0
pycparser==2.19
pyowm==2.10.0
python-dotenv==0.11.0
python-telegram-bot==12.4.2
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
tornado==6.0.3
Unidecode==1.1.1
urllib3==1.25.8
wikipedia==1.4.0
